NEW POLYFILLS AND LINK TO POKEMON API
Added fetch and promise polyfills to ensure backwards browser functionality

Added functions which load a list of pokemon from a URL into the Repository.

Created a function which displays the list in the appended button format and also within the console log

Created a function which loads the details of each pokemon into the Pokemon List.

Created a function which shows the details of the selected pokemon in the console log on click