MODAL AND STYLING ADDED
Code for creating a modal added to the pokemon api code including html and css classes and containers

For a modal to be linked to the show details function which listens for a button click.

Creation of elements which allow for the api to be queried and properties to be appended to the modal

for the types property to be part of a loop which extracts all of the types.

commenting added to the function to explain purpose

css styling changes made